Changed int to long long for number of rows

+++ Week3/Q3MergeTables/Q3MergeTables/DisjointSet.h
@@ -2,17 +2,17 @@
 #include <vector>;
 
 class DisjointSet {
-	int _max;
+	long long _max;
 	std::vector<int> _parent;
 	std::vector<int> _rank;
-	std::vector<int> _rows;	// or Tables with rows
+	std::vector<long long> _rows;	// or Tables with rows
 	std::vector<int> _result;
 	int Find(const int) const;
 	int GetParent(const int);
 	void PathCompression(const std::vector<int>&, const int);
 
 public:
-	DisjointSet(const std::vector<int>&, const int);
+	DisjointSet(const std::vector<long long>&, const long long);
 	void Merge(const int, const int);
 	void PrintResult() const;
 };
